+++ core/android/analytics/src/main/AndroidManifest.xml
