+++ core/android/repository/build.gradle.kts
@@ -14,7 +14,7 @@ dependencies {
     implementation(projects.core.common.network)
     implementation(projects.core.common.notif)
     implementation(projects.core.common.time)
-    implementation(projects.core.common.tracking)
+    implementation(projects.core.common.analytics)
 }
 
 android {
