+++ core/android/viewmodel/build.gradle.kts
@@ -7,10 +7,12 @@ plugins {
 dependencies {
     api(libs.androidx.lifecycle.viewModelCompose)
 
+    api(projects.core.common.analytics)
     api(projects.core.common.coroutines)
     api(projects.core.common.debug)
     api(projects.core.common.list)
     api(projects.core.common.nav)
+    api(projects.core.common.perf)
     api(projects.core.common.repository)
     api(projects.core.common.urlinfo)
