+++ core/common/analytics/build.gradle.kts
