+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/FrameInfo.kt
@@ -1,0 +1,7 @@
+package com.vgleadsheets.perf.common
+
+data class FrameInfo(
+    val startTimeNanos: Long,
+    val durationOnUiThreadNanos: Long,
+    val isJank: Boolean
+)
