+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/FrameTimeStats.kt
@@ -1,0 +1,9 @@
+package com.vgleadsheets.perf.common
+
+data class FrameTimeStats(
+    val jankFrames: Int,
+    val totalFrames: Int,
+    val medianMillis: Long,
+    val ninetyFiveMillis: Long,
+    val ninetyNineMillis: Long
+)
