+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/InvalidateInfo.kt
@@ -1,0 +1,6 @@
+package com.vgleadsheets.perf.common
+
+data class InvalidateInfo(
+    val startTimeNanos: Long,
+    val durationNanos: Long
+)
