+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/InvalidateStats.kt
@@ -1,0 +1,10 @@
+package com.vgleadsheets.perf.common
+
+data class InvalidateStats(
+    val jankInvalidates: Int,
+    val totalInvalidates: Int,
+    val totalInvalidateTimeMillis: Long,
+    val medianMillis: Long,
+    val ninetyFiveMillis: Long,
+    val ninetyNineMillis: Long
+)
