+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/PerfBackend.kt
@@ -1,0 +1,8 @@
+package com.vgleadsheets.perf.common
+
+interface PerfBackend {
+    fun startScreen(screenName: String)
+    fun finishTrace(screenName: String, perfStage: PerfStage)
+    fun cancel(screenName: String)
+    fun error(message: String)
+}
