+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/PerfStage.kt
@@ -1,0 +1,11 @@
+package com.vgleadsheets.perf.common
+
+enum class PerfStage {
+    VIEW_CREATED,
+    TITLE_LOADED,
+    TRANSITION_START,
+    PARTIAL_CONTENT_LOAD,
+    FULL_CONTENT_LOAD,
+    CANCELLATION,
+    COMPLETION
+}
