+++ core/common/perf/src/main/java/com/vgleadsheets/perf/common/ScreenLoadStatus.kt
@@ -1,0 +1,9 @@
+package com.vgleadsheets.perf.common
+
+import java.util.EnumMap
+
+data class ScreenLoadStatus(
+    val name: String,
+    val startTimeNanos: Long,
+    val stageDurationMillis: Map<PerfStage, Long?> = EnumMap(PerfStage::class.java),
+)
