--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/FrameInfo.kt
@@ -1,7 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-data class FrameInfo(
-    val startTimeNanos: Long,
-    val durationOnUiThreadNanos: Long,
-    val isJank: Boolean
-)
