--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/FrameTimeStats.kt
@@ -1,9 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-data class FrameTimeStats(
-    val jankFrames: Int,
-    val totalFrames: Int,
-    val medianMillis: Long,
-    val ninetyFiveMillis: Long,
-    val ninetyNineMillis: Long
-)
