--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/InvalidateInfo.kt
@@ -1,6 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-data class InvalidateInfo(
-    val startTimeNanos: Long,
-    val durationNanos: Long
-)
