--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/InvalidateStats.kt
@@ -1,10 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-data class InvalidateStats(
-    val jankInvalidates: Int,
-    val totalInvalidates: Int,
-    val totalInvalidateTimeMillis: Long,
-    val medianMillis: Long,
-    val ninetyFiveMillis: Long,
-    val ninetyNineMillis: Long
-)
