--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/PerfStage.kt
@@ -1,11 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-enum class PerfStage {
-    VIEW_CREATED,
-    TITLE_LOADED,
-    TRANSITION_START,
-    PARTIAL_CONTENT_LOAD,
-    FULL_CONTENT_LOAD,
-    CANCELLATION,
-    COMPLETION
-}
