--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/PerfTrackingBackend.kt
@@ -1,8 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-interface PerfTrackingBackend {
-    fun startScreen(screenName: String)
-    fun finishTrace(screenName: String, perfStage: PerfStage)
-    fun cancel(screenName: String)
-    fun error(message: String)
-}
