--- core/common/perf/src/main/java/com/vgleadsheets/perf/tracking/common/ScreenLoadStatus.kt
@@ -1,9 +1,0 @@
-package com.vgleadsheets.perf.tracking.common
-
-import java.util.EnumMap
-
-data class ScreenLoadStatus(
-    val name: String,
-    val startTimeNanos: Long,
-    val stageDurationMillis: Map<PerfStage, Long?> = EnumMap(PerfStage::class.java),
-)
