+++ core/common/repository/build.gradle.kts
@@ -12,6 +12,6 @@ dependencies {
     implementation(projects.core.common.network)
     implementation(projects.core.common.notif)
     implementation(projects.core.common.time)
-    implementation(projects.core.common.tracking)
+    implementation(projects.core.common.analytics)
     implementation(projects.core.common.ui.strings)
 }
