+++ core/fake/analytics/src/main/AndroidManifest.xml
