+++ core/fake/analytics/src/main/java/com/vgleadsheets/analytics/AnalyticsModule.kt
@@ -1,0 +1,16 @@
+package com.vgleadsheets.analytics
+
+import com.vgleadsheets.logging.Hatchet
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object AnalyticsModule {
+    @Provides
+    @Singleton
+    fun provideAnalytics(hatchet: Hatchet): Analytics = NoopAnalytics(hatchet)
+}
