+++ core/fake/perf/src/main/java/com/vgleadsheets/perf/noop/NoopBackend.kt
@@ -1,0 +1,17 @@
+package com.vgleadsheets.perf.noop
+
+import com.vgleadsheets.perf.common.PerfStage
+import com.vgleadsheets.perf.common.PerfBackend
+
+class NoopBackend : PerfBackend {
+    override fun startScreen(screenName: String) = Unit
+
+    override fun finishTrace(screenName: String, perfStage: PerfStage) = Unit
+
+    override fun cancel(screenName: String) = Unit
+
+    override fun error(message: String) {
+        // TODO Re-enable after compose refactor is done
+        // throw IllegalStateException(message)
+    }
+}
