--- core/fake/perf/src/main/java/com/vgleadsheets/perf/tracking/noop/NoopTrackingBackend.kt
@@ -1,16 +1,0 @@
-package com.vgleadsheets.perf.tracking.noop
-
-import com.vgleadsheets.perf.tracking.common.PerfStage
-import com.vgleadsheets.perf.tracking.common.PerfTrackingBackend
-
-class NoopTrackingBackend : PerfTrackingBackend {
-    override fun startScreen(screenName: String) = Unit
-
-    override fun finishTrace(screenName: String, perfStage: PerfStage) = Unit
-
-    override fun cancel(screenName: String) = Unit
-
-    override fun error(message: String) {
-        throw IllegalStateException(message)
-    }
-}
