--- core/fake/tracking/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    alias(libs.plugins.vgls.core.jvm)
-    alias(libs.plugins.vgls.di.jvm)
-}
-
-dependencies {
-    // Contains the Tracker interface
-    api(projects.core.common.tracking)
-
-    implementation(projects.core.common.coroutines)
-    implementation(projects.core.common.logging)
-}
