--- core/fake/tracking/src/main/java/com/vgleadsheets/tracking/TrackerModule.kt
@@ -1,16 +1,0 @@
-package com.vgleadsheets.tracking
-
-import com.vgleadsheets.logging.Hatchet
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object TrackerModule {
-    @Provides
-    @Singleton
-    fun provideTracker(hatchet: Hatchet): Tracker = NoopTracker(hatchet)
-}
